--- src/lib/generator/optionConfigsTypes/string.ts
@@ -1,11 +1,0 @@
-import { assertType } from '../../utils/assertType';
-import {trim} from '../defaultValues';
-
-export const stringTypeCase = (optionConfigs: any, defaultConfiguredOption: any) => {
-  if (optionConfigs.hasOwnProperty('trim')) {
-    assertType(optionConfigs.trim, 'boolean', 'trim property');
-    defaultConfiguredOption.trim = optionConfigs.trim;
-  } else {
-    defaultConfiguredOption.trim = trim;
-  }
-};
